% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secret_pw_get.R
\name{secret_pw_get}
\alias{secret_pw_get}
\title{Extract password in raw format from an environment object}
\usage{
secret_pw_get(env_name)
}
\arguments{
\item{env_name}{Name of the password variable in either your local environment or rsconnect. Passed as a string.}
}
\description{
Extract password in raw format from an environment object
}
