% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secret_read.R
\name{secret_read}
\alias{secret_read}
\title{Read in password protected json file from the folder inst/secret}
\usage{
secret_read(path)
}
\arguments{
\item{path}{File name of the encrypted json key}
}
\description{
Read in password protected json file from the folder inst/secret
}
