% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secret_write.R
\name{secret_write}
\alias{secret_write}
\title{Write out password protected json file to the folder inst/secret}
\usage{
secret_write(name, input, dir = getwd())
}
\arguments{
\item{input}{File name of the original unencrypted json key}

\item{dir}{Location to save the inst/secret folder to. Defaults to the current working directory}

\item{path}{File name of the encrypted json key}
}
\description{
Write out password protected json file to the folder inst/secret
}
